requirements.txt
appdirs>=1.4.4
asgiref>=3.5.0
attrs>=21.4.0
beautifulsoup4>=4.11.1
cattrs>=1.10.0
certifi>=2021.10.8
cffi>=1.15.0
charset-normalizer>=2.0.12
colorama>=0.4.4
dj-database-url>=0.5.0
dj-static>=0.0.6
Django>=4.0.4
django-bootstrap-v5>=1.0.11
django-htmlmin>=0.11.0
django-jazzmin>=2.5.0
django-clearcache>=1.2.1
forex-python>=1.8
gunicorn>=20.1.0
idna>=3.3
install>=1.3.5
msgpack>=1.0.3
Pillow>=9.1.0
psycopg2>=2.9.3
psycopg2-binary>=2.9.3
py-algorand-sdk>=1.11.0
pycparser>=2.21
pycryptodomex>=3.14.1
PyNaCl>=1.5.0
python-coinmarketcap>=0.3
python-dateutil>=1.5
python-memcached>=1.59.0
qrcode>=7.3.1
ratelimit>=2.2.1
requests>=2.27.1
requests-cache>=0.9.3
simplejson>=3.17.6
six>=1.16.0
tzlocal>=4.2
soupsieve>=2.3.2.post1
sqlparse>=0.4.2
static3>=0.7.0
tzdata>=2022.1
url-normalize>=1.4.3
urllib3>=1.26.9
whitenoise>=6.0.0
DateTime>=4.4
clear-cache>=1.0.0
django-clear-cache>=0.3.0
pyshorteners>=1.0.1
django-apscheduler>=0.6.2
djangorestframework>=3.13.1
django-cleanup>=6.0.0